moving SymbolicHamiltonian to core
