+++ src/qibo/abstractions/hamiltonians.py
@@ -136,185 +136,6 @@ class Hamiltonian(ABC):
         raise_error(NotImplementedError)
 
 
-class SymbolicHamiltonian(ABC):
-    """Parses symbolic Hamiltonians defined using ``sympy``.
-
-    This class should not be used by users.
-    It is used internally to help creating
-    :class:`qibo.abstractions.hamiltonians.Hamiltonian` and
-    :class:`qibo.abstractions.hamiltonians.TrotterHamiltonian` objects for Hamiltonians
-    defined using symbols. For more information we refer to the
-    :meth:`qibo.abstractions.hamiltonians.Hamiltonian.from_symbolic`
-    and :meth:`qibo.abstractions.hamiltonians.TrotterHamiltonian.from_symbolic` methods.
-
-    Args:
-        symbolic_hamiltonian (sympy.Expr): The full Hamiltonian written with
-            symbols.
-        symbol_map (dict): Dictionary that maps each symbol to a pair of
-            (target, matrix).
-    """
-    import sympy
-    from qibo import matrices
-
-    def __init__(self, hamiltonian, symbol_map):
-        if not issubclass(hamiltonian.__class__, self.sympy.Expr):
-            raise_error(TypeError, "Symbolic Hamiltonian should be a `sympy` "
-                                   "expression but is {}."
-                                   "".format(type(hamiltonian)))
-        if not isinstance(symbol_map, dict):
-            raise_error(TypeError, "Symbol map must be a dictionary but is "
-                                   "{}.".format(type(symbol_map)))
-        for k, v in symbol_map.items():
-            if not isinstance(k, self.sympy.Symbol):
-                raise_error(TypeError, "Symbol map keys must be `sympy.Symbol` "
-                                       "but {} was found.".format(type(k)))
-            if not isinstance(v, tuple):
-                raise_error(TypeError, "Symbol map values must be tuples but "
-                                       "{} was found.".format(type(v)))
-            if len(v) != 2:
-                raise_error(ValueError, "Symbol map values must be tuples of "
-                                        "length 2 but length {} was found."
-                                        "".format(len(v)))
-        self.symbolic = self.sympy.expand(hamiltonian)
-        self.map = symbol_map
-
-        term_dict = self.symbolic.as_coefficients_dict()
-        self.constant = 0
-        if 1 in term_dict:
-            self.constant = self.matrices.dtype(term_dict.pop(1))
-        self.terms = dict()
-        target_ids = set()
-        for term, coeff in term_dict.items():
-            targets, matrices = [], [self.matrices.dtype(coeff)]
-            for factor in term.as_ordered_factors():
-                if factor.is_symbol:
-                    self._check_symbolmap(factor)
-                    itarget = self.map[factor][0]
-                    ivalues = self.map[factor][1]
-                    if hasattr(ivalues, '__len__'):
-                        targets.append(itarget)
-                        matrices.append(ivalues)
-                    else:
-                        matrices[0] *= ivalues
-                elif isinstance(factor, self.sympy.Pow):
-                    base, pow = factor.args
-                    assert isinstance(pow, self.sympy.Integer)
-                    self._check_symbolmap(base)
-                    targets.append(self.map[base][0])
-                    matrix = self.map[base][1]
-                    for _ in range(int(pow) - 1):
-                        matrix = matrix.dot(matrix)
-                    matrices.append(matrix)
-                else:
-                    raise_error(ValueError, f"Cannot parse factor {factor}.")
-            target_ids |= set(targets)
-            targets, matrices = tuple(targets), tuple(matrices)
-            if targets in self.terms:
-                self.terms[targets] += matrices
-            else:
-                self.terms[targets] = matrices
-        self.nqubits = max(target_ids) + 1
-
-    def _check_symbolmap(self, s):
-        """Checks if symbol exists in the given symbol map."""
-        if s not in self.map:
-            raise_error(ValueError, f"Symbolic Hamiltonian contains symbol {s} "
-                                    "which does not exist in the symbol map.")
-
-    @staticmethod
-    @abstractmethod
-    def multikron(matrix_list): # pragma: no cover
-        """Calculates Kronecker product of a list of matrices.
-
-        Args:
-            matrices (list): List of matrices as ``np.ndarray``s.
-
-        Returns:
-            ``np.ndarray`` of the Kronecker product of all ``matrices``.
-        """
-        raise_error(NotImplementedError)
-
-    def full_matrices(self):
-        """Generator of matrices for each symbolic Hamiltonian term.
-
-        Returns:
-            Matrices of shape ``(2 ** nqubits, 2 ** nqubits)`` for each term in
-            the given symbolic form. Here ``nqubits`` is the total number of
-            qubits that the Hamiltonian acts on.
-        """
-        for targets, matrices in self.terms.items():
-            matrix_list = self.nqubits * [self.matrices.I]
-            n = len(targets)
-            total = 0
-            for i in range(0, len(matrices), n + 1):
-                for t, m in zip(targets, matrices[i + 1: i + n + 1]):
-                    matrix_list[t] = m
-                total += matrices[i] * self.multikron(matrix_list)
-            yield total
-
-    def partial_matrices(self):
-        """Generator of matrices for each symbolic Hamiltonian term.
-
-        Returns:
-            Matrices of shape ``(2 ** ntargets, 2 ** ntargets)`` for each term
-            in the given symbolic form. Here ``ntargets`` is the number of
-            qubits that the corresponding term acts on.
-        """
-        for targets, matrices in self.terms.items():
-            n = len(targets)
-            matrix = 0
-            for i in range(0, len(matrices), n + 1):
-                matrix += matrices[i] * self.multikron(
-                    matrices[i + 1: i + n + 1])
-            yield targets, matrix
-
-    @abstractmethod
-    def dense_matrix(self): # pragma: no cover
-        """Creates the full Hamiltonian matrix.
-
-        Useful for creating :class:`qibo.abstractions.hamiltonians.Hamiltonian`
-        object equivalent to the given symbolic Hamiltonian.
-
-        Returns:
-            Full Hamiltonian matrix of shape ``(2 ** nqubits, 2 ** nqubits)``.
-        """
-        raise_error(NotImplementedError)
-
-    @abstractmethod
-    def merge_one_qubit(self, terms): # pragma: no cover
-        """Merges one-qubit matrices to the two-qubit terms for efficiency.
-
-        This works for Hamiltonians with one and two qubit terms only.
-        The two qubit terms should be sufficiently many so that every
-        qubit appears as the first target at least once.
-
-        Args:
-            terms (dict): Dictionary that maps tuples of targets to the matrix
-                          that acts on these on targets.
-
-        Returns:
-            The given ``terms`` dictionary updated so that one-qubit terms
-            are merged to two-qubit ones.
-        """
-        raise_error(NotImplementedError)
-
-    def trotter_terms(self):
-        """Creates a dictionary of targets and matrices.
-
-        Useful for creating :class:`qibo.abstractions.hamiltonians.TrotterHamiltonian`
-        objects.
-
-        Returns:
-            terms (dict): Dictionary that maps tuples of targets to the matrix
-                          that acts on these on targets.
-            constant (float): The overall constant term of the Hamiltonian.
-        """
-        terms = {t: m for t, m in self.partial_matrices()}
-        if set(len(t) for t in terms.keys()) == {1, 2}:
-            terms = self.merge_one_qubit(terms)
-        return terms, self.constant
-
-
 class TrotterHamiltonian(Hamiltonian):
     """Hamiltonian operator used for Trotterized time evolution.
